Split fields processing into separate file
